+++ changelog.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- Wrap errors in runner (#487)
+
 ## v1.0.1321 (2018-08-14)
 
 - Add OCI object store support (#465)